requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
mysql-connector-python==8.0.25
python-dotenv==1.0.0